+++ src/plone/testing/__init__.py
@@ -1,6 +1,4 @@
+# -*- coding: utf-8 -*-
 # Convenience imports
-
-from plone.testing.layer import (
-        Layer,
-        layered
-    )
+from plone.testing.layer import Layer    # noqa
+from plone.testing.layer import layered  # noqa