+++ src/plone/testing/_z2_testbrowser.py
@@ -1,14 +1,13 @@
-import sys
-import re
-import base64
-import rfc822
-import urllib
-import urllib2
-
+# -*- coding: utf-8 -*-
 from cStringIO import StringIO
-
+import base64
 import mechanize
 import pkg_resources
+import re
+import rfc822
+import sys
+import urllib
+import urllib2
 
 # Using a from-import here to avoid an AttributeError below when using
 # zope.testbrowser 4.x without zope.app.testing:
@@ -18,6 +17,7 @@ import zope.testbrowser.browser
 
 try:
     pkg_resources.get_distribution('Zope2>=2.13')
+
     def get_cookies(request):
         return request.response._cookie_list()
 except (pkg_resources.VersionConflict, pkg_resources.DistributionNotFound):
@@ -31,15 +31,30 @@ class Browser(zope.testbrowser.browser.Browser):
     """
 
     def __init__(self, app, url=None):
-        super(Browser, self).__init__(url=url, mech_browser=Zope2MechanizeBrowser(app))
+        super(
+            Browser,
+            self).__init__(
+            url=url,
+            mech_browser=Zope2MechanizeBrowser(app))
+
 
 class Zope2MechanizeBrowser(mechanize.Browser):
     """A mechanize browser class that uses the Zope 2 publisher to talk HTTP
     """
 
-    default_schemes    = ['http']
-    default_others     = ['_http_error', '_http_request_upgrade', '_http_default_error']
-    default_features   = ['_redirect', '_cookies', '_referer', '_refresh','_equiv', '_basicauth', '_digestauth' ]
+    default_schemes = ['http']
+    default_others = [
+        '_http_error',
+        '_http_request_upgrade',
+        '_http_default_error']
+    default_features = [
+        '_redirect',
+        '_cookies',
+        '_referer',
+        '_refresh',
+        '_equiv',
+        '_basicauth',
+        '_digestauth']
 
     def __init__(self, app, *args, **kws):
 
@@ -51,6 +66,7 @@ class Zope2MechanizeBrowser(mechanize.Browser):
         self.default_others = [cls for cls in self.default_others
                                if cls in mechanize.Browser.handler_classes]
         mechanize.Browser.__init__(self, *args, **kws)
+
 
 class Zope2HTTPHandler(urllib2.HTTPHandler):
     """A protocol handler that uses the Zope 2 publisher to talk HTTP
@@ -64,6 +80,7 @@ class Zope2HTTPHandler(urllib2.HTTPHandler):
         def connectionFactory(host, timeout=None):
             return Zope2Connection(self.app, host, timeout=timeout)
         return self.do_open(connectionFactory, req)
+
 
 class Zope2Connection(PublisherConnection):
     """A urllib2-compatible connection that can talk to the Zope 2 publisher.
@@ -92,9 +109,11 @@ class Zope2Connection(PublisherConnection):
                 # only change non-literal header names
                 key = "%s%s" % (key[:1].upper(), key[1:])
                 start = 0
-                l = key.find('-',start)
+                l = key.find('-', start)
                 while l >= start:
-                    key = "%s-%s%s" % (key[:l],key[l+1:l+2].upper(),key[l+2:])
+                    key = "%s-%s%s" % (key[:l],
+                                       key[l + 1:l + 2].upper(),
+                                       key[l + 2:])
                     start = l + 1
                     l = key.find('-', start)
             headers.append((key, val))
@@ -108,6 +127,7 @@ class Zope2Connection(PublisherConnection):
         content = self.response.getBody()
 
         return PublisherResponse(content, headers, status, reason)
+
 
 def saveState(func):
     """Save threadlocal state (security manager, local component site) before
@@ -128,10 +148,14 @@ def saveState(func):
     return wrapped_func
 
 HEADER_RE = re.compile('(\S+): (.+)$')
+
+
 def splitHeader(header):
     return HEADER_RE.match(header).group(1, 2)
 
 BASIC_RE = re.compile('Basic (.+)?:(.+)?$')
+
+
 def authHeader(header):
     match = BASIC_RE.match(header)
     if match:
@@ -143,6 +167,7 @@ def authHeader(header):
         auth = base64.encodestring('%s:%s' % (u, p))
         return 'Basic %s' % auth[:-1]
     return header
+
 
 class Zope2Caller(object):
     """Functional testing caller that can execute HTTP requests via the
@@ -172,7 +197,7 @@ class Zope2Caller(object):
         # Split off and parse the command line
         l = requestString.find('\n')
         commandLine = requestString[:l].rstrip()
-        requestString = requestString[l+1:]
+        requestString = requestString[l + 1:]
         method, url, protocol = commandLine.split()
 
         instream = StringIO(requestString)
@@ -189,14 +214,15 @@ class Zope2Caller(object):
         elif len(p) == 2:
             [env['PATH_INFO'], env['QUERY_STRING']] = p
         else:
-            raise TypeError, ''
+            raise TypeError('')
 
         # If you followed closely, you notice that one part of the url
         # gets unquoted (PATH_INFO) while the other (QUERY_STRING)
         # doesn't That complies with what the ZSERVER does.
         env['PATH_INFO'] = urllib.unquote(env['PATH_INFO'])
 
-        headers = [splitHeader(header) for header in rfc822.Message(instream).headers]
+        headers = [splitHeader(header)
+                   for header in rfc822.Message(instream).headers]
 
         # Store request body without headers
         instream = StringIO(instream.read())
@@ -207,7 +233,7 @@ class Zope2Caller(object):
                 name = 'HTTP_' + name
             env[name] = value.rstrip()
 
-        if env.has_key('HTTP_AUTHORIZATION'):
+        if 'HTTP_AUTHORIZATION' in env:
             env['HTTP_AUTHORIZATION'] = authHeader(env['HTTP_AUTHORIZATION'])
 
         outstream = StringIO()
@@ -218,7 +244,7 @@ class Zope2Caller(object):
                        stdin=instream,
                        environ=env,
                        debug=not handle_errors,
-                      )
+                       )
 
         self.app._p_jar.sync()