+++ src/plone/testing/security.py
@@ -1,9 +1,10 @@
+# -*- coding: utf-8 -*-
 """Security helpers and layers
 """
-
 from plone.testing import Layer
 
 _checkersStack = []
+
 
 def pushCheckers():
     """Push the current set of security checkers onto a stack. You should
@@ -17,6 +18,7 @@ def pushCheckers():
 
     _checkersStack.append(checker._checkers.copy())
 
+
 def popCheckers():
     """Pop the most recently pushed set of security checkers from the stack.
     You should normally do this during layer tear-down. You *must* keep calls
@@ -28,6 +30,7 @@ def popCheckers():
     from zope.security import checker
 
     checker._checkers = _checkersStack.pop()
+
 
 class Checkers(Layer):
     """Ensures correct isolation of security checkers in zope.security.